+++ .gitignore
@@ -10,6 +10,7 @@
 .phpunit.result.cache
 Homestead.json
 Homestead.yaml
+pnpm-lock.yaml
 auth.json
 npm-debug.log
 yarn-error.log
