+++ resources/views/.gitignore
@@ -1,1 +1,2 @@
 _throw.md
+*.md
