+++ storage/app/json/0109/2030/12.json
@@ -1,0 +1,1 @@
+{"status":true,"data":{"id":"0109","lokasi":"KAB. ACEH TENGGARA","daerah":"ACEH","koordinat":{"lat":3.484766666666667,"lon":97.81086111111111,"lintang":"3\u00b0 29' 5.16\" N","bujur":"97\u00b0 48' 39.1\" E"},"jadwal":[{"tanggal":"Minggu, 01\/12\/2030","imsak":"04:49","subuh":"04:59","terbit":"06:16","dhuha":"06:45","dzuhur":"12:21","ashar":"15:44","maghrib":"18:19","isya":"19:32","date":"2030-12-01"},{"tanggal":"Senin, 02\/12\/2030","imsak":"04:50","subuh":"05:00","terbit":"06:17","dhuha":"06:46","dzuhur":"12:22","ashar":"15:44","maghrib":"18:19","isya":"19:33","date":"2030-12-02"},{"tanggal":"Selasa, 03\/12\/2030","imsak":"04:50","subuh":"05:00","terbit":"06:17","dhuha":"06:46","dzuhur":"12:22","ashar":"15:45","maghrib":"18:20","isya":"19:33","date":"2030-12-03"},{"tanggal":"Rabu, 04\/12\/2030","imsak":"04:50","subuh":"05:00","terbit":"06:18","dhuha":"06:46","dzuhur":"12:22","ashar":"15:45","maghrib":"18:20","isya":"19:34","date":"2030-12-04"},{"tanggal":"Kamis, 05\/12\/2030","imsak":"04:51","subuh":"05:01","terbit":"06:18","dhuha":"06:47","dzuhur":"12:23","ashar":"15:45","maghrib":"18:20","isya":"19:34","date":"2030-12-05"},{"tanggal":"Jumat, 06\/12\/2030","imsak":"04:51","subuh":"05:01","terbit":"06:18","dhuha":"06:47","dzuhur":"12:23","ashar":"15:46","maghrib":"18:21","isya":"19:34","date":"2030-12-06"},{"tanggal":"Sabtu, 07\/12\/2030","imsak":"04:52","subuh":"05:02","terbit":"06:19","dhuha":"06:48","dzuhur":"12:24","ashar":"15:46","maghrib":"18:21","isya":"19:35","date":"2030-12-07"},{"tanggal":"Minggu, 08\/12\/2030","imsak":"04:52","subuh":"05:02","terbit":"06:19","dhuha":"06:48","dzuhur":"12:24","ashar":"15:47","maghrib":"18:22","isya":"19:35","date":"2030-12-08"},{"tanggal":"Senin, 09\/12\/2030","imsak":"04:52","subuh":"05:02","terbit":"06:20","dhuha":"06:49","dzuhur":"12:24","ashar":"15:47","maghrib":"18:22","isya":"19:36","date":"2030-12-09"},{"tanggal":"Selasa, 10\/12\/2030","imsak":"04:53","subuh":"05:03","terbit":"06:20","dhuha":"06:49","dzuhur":"12:25","ashar":"15:48","maghrib":"18:22","isya":"19:36","date":"2030-12-10"},{"tanggal":"Rabu, 11\/12\/2030","imsak":"04:53","subuh":"05:03","terbit":"06:21","dhuha":"06:50","dzuhur":"12:25","ashar":"15:48","maghrib":"18:23","isya":"19:37","date":"2030-12-11"},{"tanggal":"Kamis, 12\/12\/2030","imsak":"04:54","subuh":"05:04","terbit":"06:21","dhuha":"06:50","dzuhur":"12:26","ashar":"15:49","maghrib":"18:23","isya":"19:37","date":"2030-12-12"},{"tanggal":"Jumat, 13\/12\/2030","imsak":"04:54","subuh":"05:04","terbit":"06:22","dhuha":"06:51","dzuhur":"12:26","ashar":"15:49","maghrib":"18:24","isya":"19:38","date":"2030-12-13"},{"tanggal":"Sabtu, 14\/12\/2030","imsak":"04:54","subuh":"05:04","terbit":"06:22","dhuha":"06:51","dzuhur":"12:27","ashar":"15:50","maghrib":"18:24","isya":"19:38","date":"2030-12-14"},{"tanggal":"Minggu, 15\/12\/2030","imsak":"04:55","subuh":"05:05","terbit":"06:23","dhuha":"06:52","dzuhur":"12:27","ashar":"15:50","maghrib":"18:25","isya":"19:39","date":"2030-12-15"},{"tanggal":"Senin, 16\/12\/2030","imsak":"04:55","subuh":"05:05","terbit":"06:23","dhuha":"06:52","dzuhur":"12:28","ashar":"15:51","maghrib":"18:25","isya":"19:39","date":"2030-12-16"},{"tanggal":"Selasa, 17\/12\/2030","imsak":"04:56","subuh":"05:06","terbit":"06:24","dhuha":"06:53","dzuhur":"12:28","ashar":"15:51","maghrib":"18:26","isya":"19:40","date":"2030-12-17"},{"tanggal":"Rabu, 18\/12\/2030","imsak":"04:56","subuh":"05:06","terbit":"06:24","dhuha":"06:53","dzuhur":"12:29","ashar":"15:52","maghrib":"18:26","isya":"19:40","date":"2030-12-18"},{"tanggal":"Kamis, 19\/12\/2030","imsak":"04:57","subuh":"05:07","terbit":"06:25","dhuha":"06:54","dzuhur":"12:29","ashar":"15:52","maghrib":"18:27","isya":"19:41","date":"2030-12-19"},{"tanggal":"Jumat, 20\/12\/2030","imsak":"04:57","subuh":"05:07","terbit":"06:25","dhuha":"06:54","dzuhur":"12:30","ashar":"15:53","maghrib":"18:27","isya":"19:41","date":"2030-12-20"},{"tanggal":"Sabtu, 21\/12\/2030","imsak":"04:58","subuh":"05:08","terbit":"06:26","dhuha":"06:55","dzuhur":"12:30","ashar":"15:53","maghrib":"18:28","isya":"19:42","date":"2030-12-21"},{"tanggal":"Minggu, 22\/12\/2030","imsak":"04:58","subuh":"05:08","terbit":"06:26","dhuha":"06:55","dzuhur":"12:31","ashar":"15:54","maghrib":"18:28","isya":"19:42","date":"2030-12-22"},{"tanggal":"Senin, 23\/12\/2030","imsak":"04:59","subuh":"05:09","terbit":"06:27","dhuha":"06:56","dzuhur":"12:31","ashar":"15:54","maghrib":"18:29","isya":"19:43","date":"2030-12-23"},{"tanggal":"Selasa, 24\/12\/2030","imsak":"04:59","subuh":"05:09","terbit":"06:27","dhuha":"06:56","dzuhur":"12:32","ashar":"15:55","maghrib":"18:29","isya":"19:43","date":"2030-12-24"},{"tanggal":"Rabu, 25\/12\/2030","imsak":"05:00","subuh":"05:10","terbit":"06:28","dhuha":"06:57","dzuhur":"12:32","ashar":"15:55","maghrib":"18:30","isya":"19:44","date":"2030-12-25"},{"tanggal":"Kamis, 26\/12\/2030","imsak":"05:00","subuh":"05:10","terbit":"06:28","dhuha":"06:57","dzuhur":"12:33","ashar":"15:56","maghrib":"18:30","isya":"19:44","date":"2030-12-26"},{"tanggal":"Jumat, 27\/12\/2030","imsak":"05:01","subuh":"05:11","terbit":"06:29","dhuha":"06:58","dzuhur":"12:33","ashar":"15:56","maghrib":"18:31","isya":"19:45","date":"2030-12-27"},{"tanggal":"Sabtu, 28\/12\/2030","imsak":"05:01","subuh":"05:11","terbit":"06:29","dhuha":"06:58","dzuhur":"12:34","ashar":"15:57","maghrib":"18:31","isya":"19:45","date":"2030-12-28"},{"tanggal":"Minggu, 29\/12\/2030","imsak":"05:02","subuh":"05:12","terbit":"06:30","dhuha":"06:59","dzuhur":"12:34","ashar":"15:57","maghrib":"18:32","isya":"19:46","date":"2030-12-29"},{"tanggal":"Senin, 30\/12\/2030","imsak":"05:02","subuh":"05:12","terbit":"06:30","dhuha":"06:59","dzuhur":"12:35","ashar":"15:58","maghrib":"18:32","isya":"19:46","date":"2030-12-30"},{"tanggal":"Selasa, 31\/12\/2030","imsak":"05:03","subuh":"05:13","terbit":"06:31","dhuha":"07:00","dzuhur":"12:35","ashar":"15:58","maghrib":"18:33","isya":"19:47","date":"2030-12-31"}]}}
