+++ vite.config.js
@@ -1,10 +1,14 @@
-import { defineConfig } from 'vite';
-import laravel from 'laravel-vite-plugin';
+import { defineConfig } from "vite";
+import laravel from "laravel-vite-plugin";
 
 export default defineConfig({
     plugins: [
         laravel({
-            input: ['resources/css/app.css', 'resources/js/app.js'],
+            input: [
+                "resources/css/app.css",
+                "resources/js/app.js",
+                "resources/js/apis.js",
+            ],
             refresh: true,
         }),
     ],
